+++ Kernel/System/Console/Command/Maint/Stats/Generate.pm
@@ -271,7 +271,6 @@ sub Run {
             StatID       => $Self->{StatID},
             GetParam     => \%GetParam,
             UserID       => 1,
-            UserLanguage => $Self->{Language},
         ),
     };